requirements.txt
# Core dependencies
requests>=2.25.0
python-dotenv>=0.19.0

